day 6
Puzzle2 was actually harder than usual due to memory limits of Javascript! It was fun to find a way to still perform the simulation asked for. Managed to find a very simple way in the end! slow.js still contains the old method I was planning on using, but just didn't work.